+++ spatialyze/utils/__init__.py
@@ -1,5 +1,4 @@
 from . import F
-from .add_recognized_objects import add_recognized_objects
 from .bbox_to_data3d import bbox_to_data3d
 from .compute_heading import compute_heading
 from .create_transform_matrix import create_transform_matrix
@@ -10,14 +9,11 @@ from .import_tables import import_tables
 from .ingest_road import ingest_road
 from .join import join
 from .overlay_trajectory import fetch_camera_config, overlay_trajectory
-from .recognize import recognize
 from .transformation import transformation
 
 __all__ = [
     "F",
-    "add_recognized_objects",
     "compute_heading",
-    "recognize",
     "df_to_camera_config",
     "overlay_trajectory",
     "create_transform_matrix",
